+++ includes/woocommerce.php
@@ -20,6 +20,9 @@ if ( ! function_exists( 'lsx_wc_support' ) ) :
 	 */
 	function lsx_wc_support() {
 		add_theme_support( 'woocommerce' );
+		// add_theme_support( 'wc-product-gallery-zoom' );
+		add_theme_support( 'wc-product-gallery-lightbox' );
+		// add_theme_support( 'wc-product-gallery-slider' );
 	}
 
 	add_action( 'after_setup_theme', 'lsx_wc_support' );