+++ includes/woocommerce.php
@@ -22,7 +22,7 @@ if ( ! function_exists( 'lsx_wc_support' ) ) :
 		add_theme_support( 'woocommerce' );
 		add_theme_support( 'wc-product-gallery-zoom' );
 		add_theme_support( 'wc-product-gallery-lightbox' );
-		// add_theme_support( 'wc-product-gallery-slider' );
+		add_theme_support( 'wc-product-gallery-slider' );
 	}
 
 	add_action( 'after_setup_theme', 'lsx_wc_support' );