+++ changelog.md
@@ -3,6 +3,7 @@
 ### 2.0.1
 * Fix - Fixed conflict from global style affecting slick dots
 * Fix - Removed the setting that turned LSX as LSX (itself) child theme
+* Dev - Added new filter to disable the global pagination: lsx_paging_nav_disable
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned