+++ changelog.md
@@ -6,6 +6,8 @@
 * Dev - Added new filter to disable the global pagination: lsx_paging_nav_disable
 * Fix - Fixed blog/search results - Custom post type visual
 * Fix - Fixed blog/search results - Botttom image only used by LSX Blog Customizer
+* Fix - Added dynamic class to WooCommerce archive wrapper
+* Tweak - WooCommerce mobile footer bar
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned