+++ changelog.md
@@ -11,6 +11,7 @@
 * Fix - Fixed <main> close tag in WooCommerce archive
 * Fix - Improve menu cart item visual
 * Fix - Improve WooCommerce visual experience
+* Tweak - New button class (red colours) for top menu: cta-red
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned