+++ changelog.md
@@ -12,8 +12,9 @@
 * Fix - Improve menu cart item visual
 * Fix - Improve WooCommerce visual experience
 * Tweak - New button class (red colours) for top menu: cta-red
-* Tweak - New hooks to allow LSX Customizer to change WooCommerce cart menu item position
-* Tweak - New hooks to allow LSX Customizer to change WooCommerce cart menu item style
+* Tweak - Small changes to allow LSX Customizer to change WooCommerce cart menu item position
+* Tweak - Small changes to allow LSX Customizer to change WooCommerce cart menu item style
+* Tweak - Small changes to allow LSX Customizer to change WooCommerce My Account menu item
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned