+++ changelog.md
@@ -17,6 +17,7 @@
 * Tweak - Small changes to allow LSX Customizer to change WooCommerce My Account menu item
 * Tweak - Add WooCommerce Wishlist compatibility
 * Tweak - Add WooCommerce Smart Coupon compatibility
+* Tweak - Add WooCommerce image lightbox effect compatibility
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned