+++ changelog.md
@@ -20,6 +20,7 @@
 * Tweak - Add WooCommerce image lightbox effect compatibility
 * Fix - Bootstrap modal z-index
 * Tweak - Add WooCommerce image slider effect compatibility
+* Tweak - Add WooCommerce Advanced Product Labels compatibility
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned